JMTools Package Documentation

Index of functions

JMTools.JMTools

JMTools.current_misr_prdct_version

JMTools.is_valid_misr_block

JMTools.is_valid_misr_camera

JMTools.is_valid_misr_orbit

JMTools.is_valid_misr_path

JMTools.is_valid_misr_prdct

JMTools.is_valid_misr_site

JMTools.is_valid_misrhr_prdct

JMTools.mk_misr_fname

JMTools.mk_misr_fpath

JMTools.mk_misr_fspec

Main module

JMTools.JMTools — Module

using JMTools

Purpose:

Provide access to MISR Tools functions.

Licensing:

Mtk C Library: Copyright © 2005 California Institute of Technology, Caltech license.

Julia function: Copyright © 2023 Michel M. Verstraete, MIT license.

Versioning:

Mtk C Library: Version 1.5.

JMtk15 package: Version 0.1.0 (2023-02-15).

Julia function: Version 0.1.0 (2023-05-15).

Note(s): N/A.

Example:

julia> using JMTools
[ Info: Precompiling JMTools [7ad2ddfe-0423-4906-9772-0d2d9ace2845]

Functions

JMTools.current_misr_prdct_version — Function

misr_prdct_version = current_misr_prdct_version(misr_prdct)

Purpose:

Return the current version label for the specified MISR product acronym.

Required positional argument(s):

misr_prdct::AbstractString: The MISR product acronym name.

Optional keyword(s): N/A.

Returned value(s):

misr_prdct_version::AbstractString: The current version label for the specified MISR product acronym.

Licensing:

Mtk C Library: Copyright © 2005 California Institute of Technology, Caltech license.

Julia function: Copyright © 2023 Michel M. Verstraete, MIT license.

Versioning:

Mtk C Library: Version 1.5.

Julia function: Version 0.1.0 (2023-05-15).

Note(s):

The version identifiers defined below may need to be updated when new products are generated at NASA ASDC.

Example(s):

julia> using JMTools

julia> misr_prdct_version = current_misr_prdct_version("L2AERO")
"F13_0023"

julia> misr_prdct_version = current_misr_prdct_version("ABcd")
ERROR: current_misr_prdct_version: Unrecognized MISR product.

JMTools.is_valid_misr_block — Function

bool, misr_block_string = is_valid_misr_block(misr_block)

Purpose:

Determine whether the specified misr_block is a valid MISR Block number, and return the string version of the MISR Path number prepended with the letter "B".

Required positional argument(s):

misr_block::Integer: The MISR Block number.

Optional keyword(s): N/A.

Returned value(s):

bool::Bool: Whether misr_block is valid or not.

misr_block_string::AbstractString: The string version of the MISR Block number.

Licensing:

Mtk C Library: Copyright © 2005 California Institute of Technology, Caltech license.

Julia function: Copyright © 2023 Michel M. Verstraete, MIT license.

Versioning:

Mtk C Library: Version 1.5.

Julia function: Version 0.1.0 (2023-05-15).

Note(s): N/A.

Example(s):

julia> using JMTools

julia> bool, misr_block_string = is_valid_misr_block(110)
(true, "B110")

julia> bool, misr_block_string = is_valid_misr_block(190)
(false, "")

JMTools.is_valid_misr_camera — Function

bool = is_valid_misr_camera(misr_camera)

Purpose:

Determine whether the specified misr_camera is a valid MISR camera name.

Required positional argument(s):

misr_camera::AbstractString: The MISR camera name.

Optional keyword(s): N/A.

Returned value(s):

bool::Bool: Whether misr_camera is valid or not.

Licensing:

Mtk C Library: Copyright © 2005 California Institute of Technology, Caltech license.

Julia function: Copyright © 2023 Michel M. Verstraete, MIT license.

Versioning:

Mtk C Library: Version 1.5.

Julia function: Version 0.1.0 (2023-05-15).

Note(s):

The input positional argument misr_camera can be specified as "*" to refer to all 9 camer names.

Example(s):

julia> using JMTools

julia> bool = is_valid_misr_camera("AA")
true

julia> bool = is_valid_misr_camera("df")
false

julia> bool = is_valid_misr_camera("*")
true

JMTools.is_valid_misr_orbit — Function

bool, misr_orbit_string = is_valid_misr_orbit(misr_orbit)

Purpose:

Determine whether the specified misr_orbit is a valid MISR Orbit number, and return the string version of the MISR Path number prepended with the letter "O".

Required positional argument(s):

misr_orbit::Integer: The MISR Orbit number.

Optional keyword(s): N/A.

Returned value(s):

bool::Bool: Whether misr_orbit is valid or not.

misr_orbit_string::AbstractString: The string version of the MISR Orbit number.

Licensing:

Mtk C Library: Copyright © 2005 California Institute of Technology, Caltech license.

Julia function: Copyright © 2023 Michel M. Verstraete, MIT license.

Versioning:

Mtk C Library: Version 1.5.

Julia function: Version 0.1.0 (2023-05-15).

Note(s): N/A.

Example(s):

julia> using JMTools

julia> bool, misr_orbit_string = is_valid_misr_orbit(68050)
(true, "O068050")

julia> bool, misr_orbit_string = is_valid_misr_orbit(994)
(false, "")

JMTools.is_valid_misr_path — Function

bool, misr_path_string = is_valid_misr_path(misr_path)

Purpose:

Determine whether the specified misr_path is a valid MISR Path number, and return the string version of the MISR Path number prepended with the letter "P".

Required positional argument(s):

misr_path::Integer: The MISR path number.

Optional keyword(s): N/A.

Returned value(s):

bool::Bool: Whether misr_path is valid or not.

misr_path_string::AbstractString: The string version of the MISR Path number.

Licensing:

Mtk C Library: Copyright © 2005 California Institute of Technology, Caltech license.

Julia function: Copyright © 2023 Michel M. Verstraete, MIT license.

Versioning:

Mtk C Library: Version 1.5.

Julia function: Version 0.1.0 (2023-05-15).

Note(s): N/A.

Example(s):

julia> using JMTools

julia> bool, misr_path_string = is_valid_misr_path(168)
(true, "P168")

julia> bool, misr_path_string = is_valid_misr_path(0)
(false, "")

JMTools.is_valid_misr_prdct — Function

bool, prdct_name, prdct_full_name, prdct_esdt = is_valid_misr_prdct(misr_prdct)

Purpose:

Determine whether the specified misr_prdct is a valid MISR product acronym, and return the product label used in filenames, the full MISR product name, and the MISR product Earth Science Data Type (ESDT) useful to acquire data files on the NASA ASDC Data Pool.

Required positional argument(s):

prdct::AbstractString: The MISR product acronym.

Optional keyword(s): N/A.

Returned value(s):

bool::Bool: Whether misr_prdct acronym is valid or not.

prdct_name::AbstractString: The product label used in filenames.

prdct_full_name::AbstractString: The full name of the MISR product.

prdct_esdt::AbstractString: The Earth Science Data Type (ESDT) label of the MISR product.

Licensing:

Mtk C Library: Copyright © 2005 California Institute of Technology, Caltech license.

Julia function: Copyright © 2023 Michel M. Verstraete, MIT license.

Versioning:

Mtk C Library: Version 1.5.

Julia function: Version 0.1.0 (2023-05-15).

Note(s):

This function currently recognizes only the MISR product acronyms defined below; additional acronyms can be added as needed.

Example(s):

julia> using JMTools

julia> bool, prdct_name, prdct_full_name, prdct_esdt = is_valid_misr_prdct("L1RTGM")
(true, "GRP_TERRAIN_GM", "Terrain Projected Radiance Global Mode Product", "MI1B2E")

julia> bool, prdct_name, prdct_full_name, prdct_esdt = is_valid_misr_prdct("AGP")
(true, "AGP", "Ancillary Geographic Product", "MIANCAGP")

julia> bool, prdct_name, prdct_full_name, prdct_esdt = is_valid_misr_prdct("agp")
(false, "", "", "")

JMTools.is_valid_misr_site — Function

bool, misr_site_label = is_valid_misr_site(misr_site;
    cap = "", sep = "", sit = false, strict = false)

Purpose:

Determine whether the specified misr_site is a valid MISR site name, and return an edited version of that site name, following the instructions provided by the optional input keywords.

Required positional argument(s):

misr_site::AbstractString: The MISR site name.

Optional keyword(s):

cap::AbstractString = "": The capitalization rule used to generate misr_site_label:

If cap == "" (default), the output argument misr_site_label is set (unmodified) as the input positional argument misr_site.

If cap == "txt", the output argument misr_site_label is set in all lower cases.

If cap == "Txt", the output argument misr_site_label is set in all lower cases but with the first letter of each distinct word in upper case.

If cap == "TXT", the input positional parameter misr_site is set in all upper cases.

sep::AbstractString = "": The word separator character used to separate words in misr_site_label:

If sep == "" (default), the output argument misr_site_label is set using the (unmodified) spelling of the input value misr_site.

If sep == " ", the words of the output argument misr_site_label are separated by a blank space.

If sep == "-", the words of the output argument misr_site_label are separated by a dash character.

If sep == "_", the words of the output argument misr_site_label are separated by an underscore character.

sit::Bool = false:

If sit == false (default), the output argument misr_site_label is not further modified beyond the rules described above.

If sit == true, the output argument misr_site_label is prepended with the identifier site, Site, or SITE capitalized in the same way as indicated by the keyword cap and separated with the character specified in the keyword sep.

strict::Bool = false:

If strict == false (default), the output argument misr_site_label uses whatever string value is provided by the input keyword misr_site.

If strict == true, the input positional parameter misr_site must be one of the recognized MISR Local Mode site names and the output argument misr_site_label will contain the corresponding value. An error condition is thrown otherwise.

Returned value(s):

bool::Bool: Whether misr_site is valid or not.

misr_site_label::AbstractString: The desired site label, properly formatted.

Licensing:

Mtk C Library: Copyright © 2005 California Institute of Technology, Caltech license.

Julia function: Copyright © 2023 Michel M. Verstraete, MIT license.

Versioning:

Mtk C Library: Version 1.5.

Julia function: Version 0.1.0 (2023-05-15).

Note(s):

The list recognized_misr_sites of recognized MISR Local Mode site names contains a subset of the African sites that have been acquired in Local Mode. It can be expanded as needed to include other regions, but those labels should include neither blank spaces nor dash characters.

If the input positional argument is not recognized, or if one of the keyword arguments cap or sep is invalid, the returned value bool is set to false and the misr_site_label contains an error message describing the exception condition.

Example(s):

julia> using JMTools

julia> bool, misr_site_label = is_valid_misr_site("egypt")
(true, "egypt")

julia> bool, misr_site_label = is_valid_misr_site("egypt", strict = true)
(false, "The input positional argument misr_site is not recognized.")

julia> bool, misr_site_label = is_valid_misr_site("egypt desert", strict = true)
(true, "egypt desert")

julia> bool, misr_site_label = is_valid_misr_site("egypt desert", cap = "Txt", strict = true)
(true, "Egypt Desert")

julia> bool, misr_site_label = is_valid_misr_site("egypt desert", cap = "TXT", sit = true, strict = true)
(true, "SITE EGYPT DESERT")

julia> bool, misr_site_label = is_valid_misr_site("egypt desert", cap = "TXT", sep = "_", strict = true)
(true, "EGYPT_DESERT")

julia> bool, misr_site_label = is_valid_misr_site("egypt desert", cap = "TXT", sep = "_", sit = true, strict = true)
(true, "SITE_EGYPT_DESERT")

julia> bool, misr_site_label = is_valid_misr_site("egypt desert", cap = "Text")
(false, "The input keyword cap is invalid.")

julia> bool, misr_site_label = is_valid_misr_site("eGypT Dessert", sep = "_")
(true, "eGypT_Dessert")

JMTools.is_valid_misrhr_prdct — Function

bool = is_valid_misrhr_prdct(misrhr_prdct)

Purpose:

Determine whether the specified misrhr_prdct is a valid MISR-HR product acronym.

Required positional argument(s):

misrhr_prdct::AbstractString: The MISR-HR product acronym.

Optional keyword(s): N/A.

Returned value(s):

bool::Bool: Whether misrhr_prdct is valid or not.

Licensing:

Mtk C Library: Copyright © 2005 California Institute of Technology, Caltech license.

Julia function: Copyright © 2023 Michel M. Verstraete, MIT license.

Versioning:

Mtk C Library: Version 1.5.

Julia function: Version 0.1.0 (2023-05-15).

Note(s):

This function currently recognizes the MISR-HR product acronyms defined below; additional acronyms can be added as needed.

Example(s):

julia> using JMTools

julia> bool = is_valid_misrhr_prdct("TIP")
true

julia> bool = is_valid_misrhr_prdct("l1B3")
false

JMTools.mk_misr_fname — Function

misr_fname = mk_misr_fname(misr_prdct, misr_path;
    misr_orbit, misr_camera, misr_site, misr_version)

Purpose:

Return a Vector of name(s) of the MISR data product file(s) corresponding to the specified input arguments.

Required positional argument(s):

misr_prdct::AbstractString: The MISR product acronym.

misr_path::Integer: The MISR Path number.

Optional keyword(s):

misr_orbit::Integer = 0: The MISR Orbit number.

misr_camera::AbstractString = "": The MISR Camera name.

misr_site::AbstractString = "": The MISR Local Mode site name.

misr_version::AbstractString = "": The MISR product version identifier.

Returned value(s):

misr_fname::Vector{AbstractString}: The name(s) of the MISR data file(s) correponding to the input arguments.

Licensing:

Mtk C Library: Copyright © 2005 California Institute of Technology, Caltech license.

Julia wrapper: Copyright © 2023 Michel M. Verstraete, MIT license.

Versioning:

Mtk C Library: Version 1.5.

JMtk15 package: Version 0.1.0 (2023-02-15).

JMTools package: Version 0.1.0 (2023-05-15).

Julia function: Version 0.1.0 (2023-05-15).

Note(s):

This function improves on the MISR Toolkit function MtkMakeFilename.c because it (1) requires only the needed input arguments and tests whether they are valid, (2) ignores unnecessary inputs, and (3) throws a meaningful error if a necessary input is missing or invalid.

The input keyword arguments are optional, but if they are explicitly mentioned in the call to the function, their values must also be specified: default values are only provided if the keywords are absent in the call. See the first two examples below.

This function cannot provide meaningful default values for the input keywords misr_orbit, misr_camera, and misr_site, but does use the most current misr_version label for the specified misr_prdct if that keyword is missing. Conversely, if the input keyword misr_version is specified, this function cannot verify its validity.

This function always returns a Vector of file name(s).

Example(s):

julia> using JMTools

julia> misr_fname = mk_misr_fname("AGP", 168)
"MISR_AM1_AGP_P168_F01_24.hdf"

julia> misr_fname = mk_misr_fname("AGP", 168; misr_orbit = 0, misr_camera = "", misr_site = "", misr_version = "")
"MISR_AM1_AGP_P168_F01_24_.hdf"

julia> misr_fname = mk_misr_fname("L1GMP", 168; misr_orbit = 68050)
"MISR_AM1_GP_GMP_P168_O068050_F03_0013_.hdf"

julia> misr_fname = mk_misr_fname("L1RCCM", 168; misr_orbit = 68050, misr_camera = "CF")
"MISR_AM1_GRP_RCCM_GM_P168_O068050_CF_F04_0025_.hdf"

julia> misr_orbit = 68050
68050

julia> misr_camera = "CF"
"CF"

julia> misr_fname = mk_misr_fname("L1RCCM", 168; misr_orbit = misr_orbit, misr_camera = misr_camera)
"MISR_AM1_GRP_RCCM_GM_P168_O068050_CF_F04_0025.hdf"

julia> misr_fname = mk_misr_fname("L1RTLM", 168; misr_orbit = 68050, misr_camera = "*", misr_site = "skukuza")
"MISR_AM1_GRP_TERRAIN_LM_P168_O068050_*_SITE_SKUKUZA_F03_0024.hdf"

julia> misr_fname = mk_misr_fname("L1GMP", 168; misr_orbit = 680500)
ERROR: mk_misr_fname: Invalid MISR Orbit number: must be in [995, 200000].

julia> misr_fname = mk_misr_fname("L1RCCM", 168; misr_orbit = 68050)
ERROR: mk_misr_fname: Missing MISR Camera name.

JMTools.mk_misr_fpath — Function

misr_fpath = mk_misr_fpath(misr_prdct, misr_path)

Purpose:

Generate the recommended directory name of the folder containing the MISR data file(s) represented by the acronym misr_prdct for the MISR PATH misr_path.

Required positional argument(s):

misr_prdct::AbstractString: The MISR product acronym.

misr_path::Integer: The MISR Path number.

Optional keyword(s): None.

Returned value(s):

misr_fpath::AbstractString: The name of the folder containing MISR data file(s) for the specified product and Path number.

Licensing:

Mtk C Library: Copyright © 2005 California Institute of Technology, Caltech license.

Julia wrapper: Copyright © 2023 Michel M. Verstraete, MIT license.

Versioning:

Mtk C Library: Version 1.5.

JMtk15 package: Version 0.1.0 (2023-02-15).

JMTools package: Version 0.1.0 (2023-05-15).

Julia function: Version 0.1.0 (2023-05-15).

Note(s):

This function assumes that MISR data file(s) are organized as described in the User Manual.

Example(s):

julia> using JMTools

julia> misr_prdct = "L1RTGM"
"L1RTGM"

julia> misr_path = 168
168

julia> misr_fpath = mk_misr_fpath(misr_prdct, misr_path)
"/Volumes/MISR_Data0/P168/L1RTGM/"

JMTools.mk_misr_fspec — Function

misr_fspec = mk_misr_fspec(misr_prdct, misr_path;
    misr_orbit = 0, misr_camera = "", misr_site = "", misr_version = "")

Purpose:

Return the specification (path and name) of the MISR data product file(s) corresponding to the specified input arguments.

Required positional argument(s):

misr_prdct::AbstractString: The MISR product acronym.

misr_path::Integer: The MISR Path number.

Optional keyword(s):

misr_orbit::Integer = 0: The MISR Orbit number.

misr_camera::AbstractString = "": The MISR Camera name.

misr_site::AbstractString = "": The MISR Local Mode site name.

misr_version::AbstractString = "": The MISR product version identifier.

Returned value(s):

misr_fspec::Vector{AbstractString}: The specification(s) (path(s) and name(s)) of the MISR data file(s) correponding to the input arguments.

Licensing:

Mtk C Library: Copyright © 2005 California Institute of Technology, Caltech license.

Julia wrapper: Copyright © 2023 Michel M. Verstraete, MIT license.

Versioning:

Mtk C Library: Version 1.5.

JMtk15 package: Version 0.1.0 (2023-02-15).

JMTools package: Version 0.1.0 (2023-05-15).

Julia function: Version 0.1.0 (2023-05-15).

Note(s):

This function assumes that MISR data file(s) are organized as described in the User Manual.

Example(s):

julia> using JMTools

julia> misr_fspec = mk_misr_fspec("L1RTGM", 168;
           misr_orbit = 68050, misr_camera = "AN", misr_site = "", misr_version = "")
"/Volumes/MISR_Data0/P168/L1RTGM/MISR_AM1_GRP_TERRAIN_GM_P168_O068050_AN_F03_0024.hdf"

julia> isfile(misr_fspec)
true